added people and payment taken update

+++ Entities/Chitti.cs
@@ -13,9 +13,9 @@ namespace CTrackAPI.Entities
 
         public string Name { get; set; }
 
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
-        public double Commission { get; set; }
+        public decimal Commission { get; set; }
 
         public int NoOfMonths { get; set; }
